🐛 [Fix] : Comment 와 Review 컴포넌트 분리
